+++ life_expectancy/tests/test_clean_data.py
@@ -21,9 +21,9 @@ def test_main_csv(monkeypatch : pytest.MonkeyPatch, data_cleaned_csv: pd.DataFra
         data_cleaned_csv (Fixture): Data fixture with cleaned data
     """
     mock_print = Mock(return_value='MonkeyPatch save')
-    monkeypatch.setattr("life_expectancy.ETLs.csv_etl.BaseETL.save_data", lambda _, save_location:
+    monkeypatch.setattr("life_expectancy.etls.csv_etl.BaseETL.save_data", lambda _, save_location:
                 print(mock_print()))
-    monkeypatch.setattr("life_expectancy.ETLs.csv_etl.CsvETL.load_data", lambda _ :
+    monkeypatch.setattr("life_expectancy.etls.csv_etl.CsvETL.load_data", lambda _ :
                 pd.read_csv(EU_FILE_NAME, ","))
     data = main(country=Country.PT).reset_index(drop=True)
     assert pd.DataFrame.equals(data, data_cleaned_csv)
@@ -40,9 +40,9 @@ def test_main_json(monkeypatch : pytest.MonkeyPatch, data_cleaned_json: pd.DataF
     """
 
     mock_print = Mock(return_value='MonkeyPatch save')
-    monkeypatch.setattr("life_expectancy.ETLs.json_etl.BaseETL.save_data", lambda _, save_location:
+    monkeypatch.setattr("life_expectancy.etls.json_etl.BaseETL.save_data", lambda _, save_location:
                 print(mock_print()))
-    monkeypatch.setattr("life_expectancy.ETLs.json_etl.JsonETL.load_data", lambda _ :
+    monkeypatch.setattr("life_expectancy.etls.json_etl.JsonETL.load_data", lambda _ :
                 pd.read_csv(EU_FILE_NAME_JSON, ","))
     data = main(etl_type = ETLSelection.JSON, country=Country.PT).reset_index(drop=True)
     assert pd.DataFrame.equals(data, data_cleaned_json)
